Move filtering logic into SchedulesTable
